Split TrackRow and DRY AccentRow
Extract shared components and utilities from TrackRow
(617 lines) and AccentRow (343 lines):

- TrackEndBar.tsx (195 lines): draggable length handle
  + free-run indicator, shared by both rows
- TrackNameButton.tsx (133 lines): name button with
  popover menu (was inline in TrackRow)
- TrackMixer.tsx (88 lines): M/S toggles + gain/pan
  knobs
- trackUtils.ts (28 lines): computeEffectiveStep and
  formatPan pure functions

TrackRow shrinks to 305 lines, AccentRow to 218.
The duplicated drag handle, effectiveStep, and
free-run indicator code is now shared via
TrackEndBar and trackUtils.
