+++ src/app/AccentRow.tsx
@@ -1,15 +1,9 @@
 "use client";
 
-import {
-  memo, useCallback, useRef, useState,
-} from 'react';
-import {
-  LongPressEventType, useLongPress,
-} from 'use-long-press';
+import { memo, useCallback } from 'react';
 import type { TrackId } from './types';
-import {
-  LONG_PRESS_MS, ENDBAR_LONG_PRESS_CANCEL_PX,
-} from './constants';
+import { computeEffectiveStep } from './trackUtils';
+import TrackEndBar from './TrackEndBar';
 import StepButton from './StepButton';
 import Knob from './Knob';
 import Tooltip from './Tooltip';
@@ -57,9 +51,6 @@ function AccentRowInner({
   onSetGain,
   onClearTrack,
 }: AccentRowProps) {
-  const gridRef = useRef<HTMLDivElement>(null);
-  const [isDragging, setIsDragging] = useState(false);
-
   const handleFreeRun = useCallback(
     () => onToggleFreeRun('ac'),
     [onToggleFreeRun]
@@ -81,80 +72,14 @@ function AccentRowInner({
     [onToggleStep, pageOffset]
   );
 
-  // ─── Drag handle logic ──────────────────────────
-  const endBarLongPress = useLongPress(
-    () => {
-      navigator.vibrate?.(10);
-      handleFreeRun();
-    },
-    {
-      detect: LongPressEventType.Touch,
-      threshold: LONG_PRESS_MS,
-      cancelOnMovement: ENDBAR_LONG_PRESS_CANCEL_PX,
-    }
+  const handleSetLength = useCallback(
+    (len: number) => onSetTrackLength('ac', len),
+    [onSetTrackLength]
   );
 
-  const effectiveStep =
-    currentStep >= 0
-      ? (isFreeRun ? totalSteps : currentStep)
-        % trackLength
-      : -1;
-
-  const lengthFromPointer = useCallback(
-    (clientX: number): number => {
-      const grid = gridRef.current;
-      if (!grid) return trackLength;
-      const rect = grid.getBoundingClientRect();
-      const x = clientX - rect.left;
-      const stepWidth = rect.width / 16;
-      const raw = Math.round(x / stepWidth);
-      return Math.max(
-        1,
-        Math.min(patternLength, raw + pageOffset)
-      );
-    },
-    [patternLength, trackLength, pageOffset]
+  const effectiveStep = computeEffectiveStep(
+    currentStep, totalSteps, isFreeRun, trackLength
   );
-
-  const handlePointerDown = useCallback(
-    (e: React.PointerEvent) => {
-      if (e.button !== 0) return;
-      if (e.ctrlKey || e.metaKey) {
-        e.preventDefault();
-        handleFreeRun();
-        return;
-      }
-      e.preventDefault();
-      (e.target as HTMLElement).setPointerCapture(
-        e.pointerId
-      );
-      setIsDragging(true);
-    },
-    [handleFreeRun]
-  );
-
-  const handlePointerMove = useCallback(
-    (e: React.PointerEvent) => {
-      if (!isDragging) return;
-      const len = lengthFromPointer(e.clientX);
-      if (len !== trackLength) {
-        onSetTrackLength('ac', len);
-      }
-    },
-    [isDragging, lengthFromPointer, trackLength,
-      onSetTrackLength]
-  );
-
-  const handlePointerUp = useCallback(() => {
-    setIsDragging(false);
-  }, []);
-
-  const handleOnPage =
-    trackLength > pageOffset
-    && trackLength <= pageOffset + 16;
-  const handlePct = handleOnPage
-    ? ((trackLength - pageOffset) / 16) * 100
-    : 0;
 
   return (
     <div>
@@ -179,7 +104,10 @@ function AccentRowInner({
           AC
         </button>
         <div className="ml-auto">
-          <Tooltip tooltipKey="accentIntensity" position="bottom">
+          <Tooltip
+            tooltipKey="accentIntensity"
+            position="bottom"
+          >
             <Knob
               value={gain}
               onChange={handleGain}
@@ -213,13 +141,16 @@ function AccentRowInner({
           >
             AC
           </button>
-          {/* Spacer matching mute + solo toggle widths */}
+          {/* Spacer matching mute + solo widths */}
           <div className="flex gap-1">
             <div className="w-6 h-6" />
             <div className="w-6 h-6" />
           </div>
           <div className="flex gap-1 ml-1">
-            <Tooltip tooltipKey="accentIntensity" position="bottom">
+            <Tooltip
+              tooltipKey="accentIntensity"
+              position="bottom"
+            >
               <Knob
                 value={gain}
                 onChange={handleGain}
@@ -233,7 +164,6 @@ function AccentRowInner({
         {/* Step grid with drag handle */}
         <div className="flex-1 relative">
           <div
-            ref={gridRef}
             data-track="ac"
             className="grid grid-cols-8 lg:grid-cols-16 gap-[3px] lg:gap-1.5"
           >
@@ -268,71 +198,16 @@ function AccentRowInner({
             )}
           </div>
 
-          {/* Draggable length handle */}
-          {handleOnPage && (
-            <div
-              role="slider"
-              aria-label="accent length"
-              aria-valuemin={1}
-              aria-valuemax={patternLength}
-              aria-valuenow={trackLength}
-              tabIndex={0}
-              onPointerDown={handlePointerDown}
-              onPointerMove={handlePointerMove}
-              onPointerUp={handlePointerUp}
-              onPointerCancel={handlePointerUp}
-              {...endBarLongPress()}
-              onContextMenu={(e) => {
-                e.preventDefault();
-                if (!isDragging) handleFreeRun();
-              }}
-              style={{
-                left: `${handlePct}%`,
-                touchAction: 'none',
-              }}
-              className={
-                'absolute top-0 h-full w-4'
-                + ' -translate-x-1/2 z-20'
-                + (isDragging
-                  ? ' cursor-col-resize'
-                  : ' cursor-default')
-                + ' before:absolute before:inset-y-0'
-                + ' before:left-1/2'
-                + ' before:-translate-x-1/2'
-                + ' before:w-1.5 before:rounded-full'
-                + ' before:transition-colors'
-                + (isDragging
-                  ? ' before:bg-neutral-300'
-                  : ' before:bg-neutral-500/60'
-                    + ' hover:before:bg-neutral-300')
-              }
-            />
-          )}
-
-          {/* Free-run indicator */}
-          {isFreeRun && handleOnPage && (
-            <span
-              aria-label="free run"
-              style={{
-                left: `${handlePct}%`,
-                fontFamily: 'var(--font-orbitron)',
-              }}
-              className={
-                'absolute top-1/2 -translate-x-1/2'
-                + ' -translate-y-1/2 z-30'
-                + ' pointer-events-none select-none'
-                + ' flex items-center'
-                + ' justify-center'
-                + ' w-3.5 h-3.5 rounded-full'
-                + ' bg-orange-500'
-                + ' text-[10px] font-bold'
-                + ' text-white'
-                + ' shadow-[0_0_6px_rgba(251,146,60,0.6)]'
-              }
-            >
-              F
-            </span>
-          )}
+          <TrackEndBar
+            trackName="accent"
+            trackLength={trackLength}
+            patternLength={patternLength}
+            pageOffset={pageOffset}
+            isFreeRun={isFreeRun}
+            onSetTrackLength={handleSetLength}
+            onToggleFreeRun={handleFreeRun}
+            showTooltip={false}
+          />
         </div>
       </div>
     </div>
