+++ src/app/TrackRow.tsx
@@ -1,155 +1,17 @@
 "use client";
 
-import {
-  memo, useCallback, useEffect, useRef, useState,
-} from 'react';
+import { memo, useCallback } from 'react';
 import type { RefObject } from 'react';
-import {
-  LongPressEventType, useLongPress,
-} from 'use-long-press';
 import type {
   StepConditions, StepLocks, TrackId,
 } from './types';
-import {
-  LONG_PRESS_MS, ENDBAR_LONG_PRESS_CANCEL_PX,
-} from './constants';
-import TrackToggle from './TrackToggle';
+import { computeEffectiveStep } from './trackUtils';
+import TrackNameButton from './TrackNameButton';
+import TrackMixer from './TrackMixer';
+import TrackEndBar from './TrackEndBar';
 import StepButton from './StepButton';
-import Knob from './Knob';
-import Tooltip from './Tooltip';
 
-function formatPan(v: number): string {
-  const pct = Math.round((v - 0.5) * 200);
-  if (pct === 0) return 'C';
-  return pct < 0 ? `L${-pct}` : `R${pct}`;
-}
-
-interface TrackNameButtonProps {
-  size: 'sm' | 'lg';
-  trackId: TrackId;
-  trackName: string;
-  isFreeRun: boolean;
-  isTriggered: boolean;
-  onToggleFreeRun: () => void;
-  onClearTrack: () => void;
-  onPlayPreview: () => void;
-}
-
-/**
- * Track name button with optional popover menu.
- * The popover (with free-run toggle) only renders
- * at 'lg' size. Owns its own menu state and refs.
- */
-function TrackNameButtonInner({
-  size,
-  trackId,
-  trackName,
-  isFreeRun,
-  isTriggered,
-  onToggleFreeRun,
-  onClearTrack,
-  onPlayPreview,
-}: TrackNameButtonProps) {
-  const [menuOpen, setMenuOpen] = useState(false);
-  const menuRef = useRef<HTMLDivElement>(null);
-  const nameRef = useRef<HTMLButtonElement>(null);
-
-  useEffect(() => {
-    if (!menuOpen) return;
-    const handleClick = (e: MouseEvent) => {
-      if (
-        menuRef.current
-        && !menuRef.current.contains(
-          e.target as Node
-        )
-        && nameRef.current
-        && !nameRef.current.contains(
-          e.target as Node
-        )
-      ) {
-        setMenuOpen(false);
-      }
-    };
-    document.addEventListener(
-      'mousedown', handleClick
-    );
-    return () =>
-      document.removeEventListener(
-        'mousedown', handleClick
-      );
-  }, [menuOpen]);
-
-  return (
-    <div className="relative">
-      <Tooltip tooltipKey={`track-${trackId}`}>
-        <button
-          ref={size === 'lg' ? nameRef : undefined}
-          onMouseDown={(e: React.MouseEvent) => {
-            if (e.button !== 0) return;
-            if (e.shiftKey) {
-              onClearTrack();
-              return;
-            }
-            if (e.metaKey || e.ctrlKey) {
-              setMenuOpen(v => !v);
-              return;
-            }
-            onPlayPreview();
-          }}
-          onTouchStart={() => onPlayPreview()}
-          onContextMenu={(e: React.MouseEvent) => {
-            e.preventDefault();
-            setMenuOpen(v => !v);
-          }}
-          className={
-            (size === 'sm'
-              ? 'text-lg'
-              : 'w-12 truncate text-xl text-left')
-            + ' font-bold uppercase tracking-wider font-[family-name:var(--font-orbitron)]'
-            + ' rounded px-1 py-0.5 transition-colors'
-            + (isTriggered
-              ? ' text-orange-300 bg-orange-400/25'
-              : isFreeRun
-                ? ' text-orange-400 bg-orange-400/10'
-                : ' text-neutral-400'
-                  + ' hover:text-neutral-200'
-                  + ' hover:bg-neutral-800/50')
-          }
-        >
-          {trackName}
-        </button>
-      </Tooltip>
-      {menuOpen && size === 'lg' && (
-        <div
-          ref={menuRef}
-          className="absolute left-0 top-full mt-1 w-36 bg-neutral-900 border border-neutral-700 rounded-lg shadow-xl z-30 overflow-hidden"
-        >
-          <button
-            onClick={() => {
-              onToggleFreeRun();
-              setMenuOpen(false);
-            }}
-            className="w-full text-left px-3 py-2 text-xs text-neutral-200 hover:bg-neutral-800 transition-colors flex items-center justify-between"
-          >
-            <span>Free-run</span>
-            <span
-              className={
-                'inline-block w-2 h-2 rounded-full '
-                + (isFreeRun
-                  ? 'bg-orange-400'
-                  : 'bg-neutral-600')
-              }
-            />
-          </button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-const TrackNameButton = memo(TrackNameButtonInner);
-
-interface TrackRowProps {
+export interface TrackRowProps {
   trackId: TrackId;
   trackName: string;
   steps: string;
@@ -201,12 +63,8 @@ interface TrackRowProps {
 }
 
 /**
- * Single track row: name, mute/solo, knob, and step
- * buttons with a draggable length handle. Steps beyond
- * the track's length are dimmed and non-interactive.
- * Each track computes its own effective running-light
- * position for polyrhythms. Clicking the track name
- * opens a popover with per-track settings.
+ * Single track row: name, mute/solo, knobs, and step
+ * buttons with a draggable length handle.
  */
 function TrackRowInner({
   trackId,
@@ -242,6 +100,7 @@ function TrackRowInner({
   onPlainClick,
   onClearSelection,
 }: TrackRowProps) {
+  // ─── Bound callbacks ──────────────────────────
   const handleMute = useCallback(
     () => onToggleMute(trackId),
     [onToggleMute, trackId]
@@ -270,88 +129,16 @@ function TrackRowInner({
     () => onPlayPreview(trackId),
     [onPlayPreview, trackId]
   );
-
-  // ─── Drag handle state ─────────────────────────
-  const gridRef = useRef<HTMLDivElement>(null);
-  const [isDragging, setIsDragging] = useState(false);
-
-  const endBarLongPress = useLongPress(
-    () => {
-      navigator.vibrate?.(10);
-      handleFreeRun();
-    },
-    {
-      detect: LongPressEventType.Touch,
-      threshold: LONG_PRESS_MS,
-      cancelOnMovement: ENDBAR_LONG_PRESS_CANCEL_PX,
-    }
+  const handleSetLength = useCallback(
+    (len: number) => onSetTrackLength(trackId, len),
+    [onSetTrackLength, trackId]
   );
 
-  const effectiveStep =
-    currentStep >= 0
-      ? (isFreeRun ? totalSteps : currentStep)
-        % trackLength
-      : -1;
-
-  // ─── Drag handle logic ──────────────────────────
-  const lengthFromPointer = useCallback(
-    (clientX: number): number => {
-      const grid = gridRef.current;
-      if (!grid) return trackLength;
-      const rect = grid.getBoundingClientRect();
-      const x = clientX - rect.left;
-      const stepWidth = rect.width / 16;
-      const raw = Math.round(x / stepWidth);
-      return Math.max(
-        1,
-        Math.min(
-          patternLength, raw + pageOffset
-        )
-      );
-    },
-    [patternLength, trackLength, pageOffset]
+  const effectiveStep = computeEffectiveStep(
+    currentStep, totalSteps, isFreeRun, trackLength
   );
 
-  const handlePointerDown = useCallback(
-    (e: React.PointerEvent) => {
-      if (e.button !== 0) return;
-      if (e.ctrlKey || e.metaKey) {
-        e.preventDefault();
-        handleFreeRun();
-        return;
-      }
-      e.preventDefault();
-      (e.target as HTMLElement).setPointerCapture(
-        e.pointerId
-      );
-      setIsDragging(true);
-    },
-    [handleFreeRun]
-  );
-
-  const handlePointerMove = useCallback(
-    (e: React.PointerEvent) => {
-      if (!isDragging) return;
-      const len = lengthFromPointer(e.clientX);
-      if (len !== trackLength) {
-        onSetTrackLength(trackId, len);
-      }
-    },
-    [isDragging, lengthFromPointer, trackLength,
-      onSetTrackLength, trackId]
-  );
-
-  const handlePointerUp = useCallback(() => {
-    setIsDragging(false);
-  }, []);
-
-  const handleOnPage =
-    trackLength > pageOffset
-    && trackLength <= pageOffset + 16;
-  const handlePct = handleOnPage
-    ? ((trackLength - pageOffset) / 16) * 100
-    : 0;
-
+  // ─── Page-offset wrappers ─────────────────────
   const handleToggleStep = useCallback(
     (tid: TrackId, localStep: number) =>
       onToggleStep(tid, localStep + pageOffset),
@@ -396,37 +183,18 @@ function TrackRowInner({
           onPlayPreview={handlePlayPreview}
         />
         <div className="flex gap-1 ml-auto items-center">
-          <TrackToggle
-            variant="mute"
-            active={isMuted}
+          <TrackMixer
             trackName={trackName}
-            size="lg"
-            onToggle={handleMute}
+            isMuted={isMuted}
+            isSolo={isSolo}
+            gain={gain}
+            pan={pan}
+            size="sm"
+            onToggleMute={handleMute}
+            onToggleSolo={handleSolo}
+            onSetGain={handleGain}
+            onSetPan={handlePan}
           />
-          <TrackToggle
-            variant="solo"
-            active={isSolo}
-            trackName={trackName}
-            size="lg"
-            onToggle={handleSolo}
-          />
-          <Knob
-            value={gain}
-            onChange={handleGain}
-            trackName={trackName}
-            size={20}
-          />
-          <Tooltip tooltipKey="pan">
-            <Knob
-              value={pan}
-              onChange={handlePan}
-              trackName={trackName}
-              size={20}
-              defaultValue={0.5}
-              ariaPrefix="Pan"
-              formatValue={formatPan}
-            />
-          </Tooltip>
         </div>
       </div>
 
@@ -443,45 +211,23 @@ function TrackRowInner({
             onClearTrack={handleClearTrack}
             onPlayPreview={handlePlayPreview}
           />
-          <div className="flex gap-1">
-            <TrackToggle
-              variant="mute"
-              active={isMuted}
-              trackName={trackName}
-              size="md"
-              onToggle={handleMute}
-            />
-            <TrackToggle
-              variant="solo"
-              active={isSolo}
-              trackName={trackName}
-              size="md"
-              onToggle={handleSolo}
-            />
-          </div>
-          <div className="flex gap-1 ml-1">
-            <Knob
-              value={gain}
-              onChange={handleGain}
-              trackName={trackName}
-            />
-            <Tooltip tooltipKey="pan">
-              <Knob
-                value={pan}
-                onChange={handlePan}
-                trackName={trackName}
-                defaultValue={0.5}
-                ariaPrefix="Pan"
-                formatValue={formatPan}
-              />
-            </Tooltip>
-          </div>
+          <TrackMixer
+            trackName={trackName}
+            isMuted={isMuted}
+            isSolo={isSolo}
+            gain={gain}
+            pan={pan}
+            size="lg"
+            onToggleMute={handleMute}
+            onToggleSolo={handleSolo}
+            onSetGain={handleGain}
+            onSetPan={handlePan}
+          />
         </div>
 
         {/* Step grid with drag handle */}
         <div className="flex-1 relative">
           <div
-            ref={gridRef}
             data-track={trackId}
             className="grid grid-cols-8 lg:grid-cols-16 gap-[3px] lg:gap-1.5"
           >
@@ -540,73 +286,15 @@ function TrackRowInner({
             )}
           </div>
 
-          {/* Draggable length handle */}
-          {handleOnPage && (
-            <Tooltip tooltipKey="lengthHandle" align="right">
-              <div
-                role="slider"
-                aria-label={`${trackName} length`}
-              aria-valuemin={1}
-              aria-valuemax={patternLength}
-              aria-valuenow={trackLength}
-              tabIndex={0}
-              onPointerDown={handlePointerDown}
-              onPointerMove={handlePointerMove}
-              onPointerUp={handlePointerUp}
-              onPointerCancel={handlePointerUp}
-              {...endBarLongPress()}
-              onContextMenu={(e) => {
-                e.preventDefault();
-                if (!isDragging) handleFreeRun();
-              }}
-              style={{
-                left: `${handlePct}%`,
-                touchAction: 'none',
-              }}
-              className={
-                'absolute top-0 h-full w-4'
-                + ' -translate-x-1/2 z-20'
-                + (isDragging
-                  ? ' cursor-col-resize'
-                  : ' cursor-default')
-                + ' before:absolute before:inset-y-0'
-                + ' before:left-1/2'
-                + ' before:-translate-x-1/2'
-                + ' before:w-1.5 before:rounded-full'
-                + ' before:transition-colors'
-                + (isDragging
-                  ? ' before:bg-neutral-300'
-                  : ' before:bg-neutral-500/60'
-                    + ' hover:before:bg-neutral-300')
-              }
-              />
-            </Tooltip>
-          )}
-
-          {/* Free-run indicator */}
-          {isFreeRun && handleOnPage && (
-            <span
-              aria-label="free run"
-              style={{
-                left: `${handlePct}%`,
-                fontFamily: 'var(--font-orbitron)',
-              }}
-              className={
-                'absolute top-1/2 -translate-x-1/2'
-                + ' -translate-y-1/2 z-30'
-                + ' pointer-events-none select-none'
-                + ' flex items-center'
-                + ' justify-center'
-                + ' w-3.5 h-3.5 rounded-full'
-                + ' bg-orange-500'
-                + ' text-[10px] font-bold'
-                + ' text-white'
-                + ' shadow-[0_0_6px_rgba(251,146,60,0.6)]'
-              }
-            >
-              F
-            </span>
-          )}
+          <TrackEndBar
+            trackName={trackName}
+            trackLength={trackLength}
+            patternLength={patternLength}
+            pageOffset={pageOffset}
+            isFreeRun={isFreeRun}
+            onSetTrackLength={handleSetLength}
+            onToggleFreeRun={handleFreeRun}
+          />
         </div>
       </div>
     </div>
